+++ packages/cache/src/bootstrap/environment.ts
@@ -1,10 +1,15 @@
 export type Base64Encoder = (input: string) => string
-export type Factory<T> = () => T
 
 export interface EnvironmentOptions {
   base64Encode: Base64Encoder
-  getHost: Factory<string>
-  getToken: Factory<string>
-  getURL: Factory<string>
+  getContext: RequestContextFactory
   userAgent?: string
+}
+
+export type RequestContextFactory = () => RequestContext
+
+export interface RequestContext {
+  host: string
+  token: string
+  url: string
 }