+++ packages/cache/src/constants.ts
@@ -1,0 +1,24 @@
+/**
+ * Number of seconds in one minute.
+ */
+export const MINUTE = 60
+
+/**
+ * Number of seconds in one hour.
+ */
+export const HOUR = 60 * MINUTE
+
+/**
+ * Number of seconds in one day.
+ */
+export const DAY = 24 * HOUR
+
+/**
+ * Number of seconds in one week.
+ */
+export const WEEK = 7 * DAY
+
+/**
+ * Number of seconds in one year.
+ */
+export const YEAR = 365 * DAY