+++ packages/cache/CHANGELOG.md
@@ -1,24 +1,78 @@
 # Changelog
 
-## [1.1.0](https://github.com/netlify/primitives/compare/cache-v1.0.0...cache@v1.1.0) (2025-02-21)
+## [1.3.0](https://github.com/netlify/primitives/compare/cache-v1.1.0...cache-v1.3.0) (2025-03-12)
 
 
 ### Features
 
-* add `@netlify/cache` module ([#32](https://github.com/netlify/primitives/issues/32)) ([e374119](https://github.com/netlify/primitives/commit/e3741190472af28a5bdcdd2e3aa6067b141cd7b5))
-* add `fetchWithCache` and `setCacheHeaders` methods ([#34](https://github.com/netlify/primitives/issues/34)) ([712b188](https://github.com/netlify/primitives/commit/712b188b9102436207fbb849cb6c20f997a522da))
-* add `getCacheStatus` utility method ([#37](https://github.com/netlify/primitives/issues/37)) ([2d739e2](https://github.com/netlify/primitives/commit/2d739e2c568aa977e5166bb532b138702e8fc33b))
+* add `getContext` method ([#77](https://github.com/netlify/primitives/issues/77)) ([ee879ae](https://github.com/netlify/primitives/commit/ee879aece706ba5e34fe1e8d46392580c4a7a7b9))
+* export time interval constants ([#81](https://github.com/netlify/primitives/issues/81)) ([60f048b](https://github.com/netlify/primitives/commit/60f048b7ba42cfee1e62725dbc67633b738d4308))
+* fix release ([0b1d397](https://github.com/netlify/primitives/commit/0b1d3975c6383b0376cbf8e4c5d2541cd9c43e4d))
+* update package.json ([6f59f75](https://github.com/netlify/primitives/commit/6f59f75851edf9ef76c254b9b3ddd33d4f1e56a3))
 
-## [1.1.0](https://github.com/netlify/primitives/compare/cache-v1.0.0...cache@v1.1.0) (2025-02-14)
+
+### Miscellaneous Chores
+
+* release 1.3.0 ([6e7237e](https://github.com/netlify/primitives/commit/6e7237ec38221382fb2d4ec49f0c184705bd68dc))
+
+## [1.3.0](https://github.com/netlify/primitives/compare/cache-v1.1.0...cache-v1.3.0) (2025-02-24)
 
 
 ### Features
 
-* add `@netlify/cache` module ([#32](https://github.com/netlify/primitives/issues/32)) ([e374119](https://github.com/netlify/primitives/commit/e3741190472af28a5bdcdd2e3aa6067b141cd7b5))
+* add `getContext` method ([#77](https://github.com/netlify/primitives/issues/77)) ([ee879ae](https://github.com/netlify/primitives/commit/ee879aece706ba5e34fe1e8d46392580c4a7a7b9))
+* fix release ([0b1d397](https://github.com/netlify/primitives/commit/0b1d3975c6383b0376cbf8e4c5d2541cd9c43e4d))
+* update package.json ([6f59f75](https://github.com/netlify/primitives/commit/6f59f75851edf9ef76c254b9b3ddd33d4f1e56a3))
 
-## [1.1.0](https://github.com/netlify/primitives/compare/cache-v1.0.0...cache@v1.1.0) (2025-02-14)
+
+### Miscellaneous Chores
+
+* release 1.3.0 ([6e7237e](https://github.com/netlify/primitives/commit/6e7237ec38221382fb2d4ec49f0c184705bd68dc))
+
+## [1.2.0](https://github.com/netlify/primitives/compare/cache-v1.1.0...cache-v1.2.0) (2025-02-22)
 
 
 ### Features
 
-* add `@netlify/cache` module ([#32](https://github.com/netlify/primitives/issues/32)) ([e374119](https://github.com/netlify/primitives/commit/e3741190472af28a5bdcdd2e3aa6067b141cd7b5))
+* fix release ([0b1d397](https://github.com/netlify/primitives/commit/0b1d3975c6383b0376cbf8e4c5d2541cd9c43e4d))
+* update package.json ([6f59f75](https://github.com/netlify/primitives/commit/6f59f75851edf9ef76c254b9b3ddd33d4f1e56a3))
+
+## [1.2.0](https://github.com/netlify/primitives/compare/cache-v1.1.0...cache-v1.2.0) (2025-02-22)
+
+
+### Features
+
+* fix release ([0b1d397](https://github.com/netlify/primitives/commit/0b1d3975c6383b0376cbf8e4c5d2541cd9c43e4d))
+* update package.json ([6f59f75](https://github.com/netlify/primitives/commit/6f59f75851edf9ef76c254b9b3ddd33d4f1e56a3))
+
+## [1.2.0](https://github.com/netlify/primitives/compare/cache-v1.1.0...cache-v1.2.0) (2025-02-22)
+
+
+### Features
+
+* fix release ([0b1d397](https://github.com/netlify/primitives/commit/0b1d3975c6383b0376cbf8e4c5d2541cd9c43e4d))
+* update package.json ([6f59f75](https://github.com/netlify/primitives/commit/6f59f75851edf9ef76c254b9b3ddd33d4f1e56a3))
+
+## [1.2.0](https://github.com/netlify/primitives/compare/cache-v1.1.0...cache-v1.2.0) (2025-02-22)
+
+
+### Features
+
+* fix release ([0b1d397](https://github.com/netlify/primitives/commit/0b1d3975c6383b0376cbf8e4c5d2541cd9c43e4d))
+* update package.json ([6f59f75](https://github.com/netlify/primitives/commit/6f59f75851edf9ef76c254b9b3ddd33d4f1e56a3))
+
+## [1.2.0](https://github.com/netlify/primitives/compare/cache-v1.1.0...cache-v1.2.0) (2025-02-22)
+
+
+### Features
+
+* fix release ([0b1d397](https://github.com/netlify/primitives/commit/0b1d3975c6383b0376cbf8e4c5d2541cd9c43e4d))
+* update package.json ([6f59f75](https://github.com/netlify/primitives/commit/6f59f75851edf9ef76c254b9b3ddd33d4f1e56a3))
+
+## [1.2.0](https://github.com/netlify/primitives/compare/cache-v1.1.0...cache-v1.2.0) (2025-02-22)
+
+
+### Features
+
+* fix release ([0b1d397](https://github.com/netlify/primitives/commit/0b1d3975c6383b0376cbf8e4c5d2541cd9c43e4d))
+* update package.json ([6f59f75](https://github.com/netlify/primitives/commit/6f59f75851edf9ef76c254b9b3ddd33d4f1e56a3))